requirements.txt
absl-py==2.2.2
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
comtypes==1.4.10
contourpy==1.3.1
cycler==0.12.1
Cython==3.1.2
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dlib==19.24.8
docx==0.2.4
easydict==1.13
executing==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
faiss-cpu==1.11.0
fastjsonschema==2.21.1
filelock==3.17.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
fsspec==2025.3.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.1
gTTS==2.5.4
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imutils==0.5.4
insightface==0.7.3
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.6.0
jaxlib==0.6.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.10.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
lxml==5.4.0
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
mistune==3.1.3
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.1
mtcnn==1.0.0
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.4.1
notebook_shim==0.2.4
numba==0.61.2
numpy==2.1.3
onnx==1.18.0
onnxruntime==1.22.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.16.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
prettytable==3.16.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
protobuf==4.25.7
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.19.1
pyparsing==3.2.1
pypiwin32==223
pyreadline3==3.5.4
pyserial==3.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-json-logger==3.3.0
pyttsx3==2.98
pytz==2025.1
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.24.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==76.0.0
shap==0.48.0
simsimd==6.5.0
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
SpeechRecognition==3.14.1
stack-data==0.6.3
stringzilla==3.12.5
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.6.11
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.88
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
Wikipedia-API==0.8.1
wrapt==1.17.2
xgboost==3.0.2
